:rocket: break apart dynamic requires for webpack
webpack statically analyzes require statements. When expressions
are found, it generates a dynamic require context based on values
it can find. However, passing a dynamic array to require is out
of scope for this simple analysis. Reworking this method makes
it possible for webpack to read and properly bundle the i18n
files for it.